Fix decimal test case and update all cases involving integer

+++ pyhive/tests/test_presto.py
@@ -42,9 +42,9 @@ class TestPresto(unittest.TestCase, DBAPITestCase):
         self.assertEqual(description, [
             ('boolean', 'boolean', None, None, None, None, True),
             ('tinyint', 'integer', None, None, None, None, True),
-            ('smallint', 'bigint', None, None, None, None, True),
-            ('int', 'bigint', None, None, None, None, True),
-            ('bigint', 'bigint', None, None, None, None, True),
+            ('smallint', 'integer', None, None, None, None, True),
+            ('int', 'integer', None, None, None, None, True),
+            ('bigint', 'integer', None, None, None, None, True),
             ('float', 'double', None, None, None, None, True),
             ('double', 'double', None, None, None, None, True),
             ('string', 'varchar', None, None, None, None, True),
